[shape_poly] linalg.lu: for shape polymorphism for native serialization on CPU.
We support polymorphism only on the batch sizes for now. The
jaxlib and C++ code support full dynamic shapes.

Also added backwards compatibility tests for the LU custom calls
for CPU, and improved the checking of LU results by checking
the invariant for the result as opposed to checking goldens.

PiperOrigin-RevId: 540464911